+++ CHANGELOG.md
@@ -110,6 +110,8 @@
 * Change behavior of "Assign" button between different lookups. Refs [UILD-745].
 * Add missing localization. Refs [UILD-784].
 * Show Authority Source name in search results. Refs [UILD-764].
+* Add support for the Geographic Coverage field. Fixes [UILD-665].
+* Change display order for titles. Refs [UILD-304].
 * Respond to the embedding application's locale config. Refs [UILD-786].
 
 [UILD-552]:https://folio-org.atlassian.net/browse/UILD-552
@@ -220,6 +222,8 @@
 [UILD-745]:https://folio-org.atlassian.net/browse/UILD-745
 [UILD-784]:https://folio-org.atlassian.net/browse/UILD-784
 [UILD-764]:https://folio-org.atlassian.net/browse/UILD-764
+[UILD-665]:https://folio-org.atlassian.net/browse/UILD-665
+[UILD-304]:https://folio-org.atlassian.net/browse/UILD-304
 [UILD-786]:https://folio-org.atlassian.net/browse/UILD-786
 
 ## 1.0.5 (2025-04-30)